Start implementation of assessment service
- create service & add to modules
- start replacing functions - first up in user dashboard
- create GQL service, unsure how we're going to use that.
