+++ src/services/gql.service.ts
@@ -1,0 +1,16 @@
+import gql from "graphql-tag";
+
+export var assessmentQuery = gql`
+query assessments($userId: String) {
+	assessments(userId: $userId) {
+	   scope
+     targetMRL
+     targetDate
+     levelSwitching
+     deskbookVersion
+     location
+     name
+	}
+}
+`
+
